refactor default head 분할 관리
