+++ src/assets/head.tsx
@@ -1,0 +1,83 @@
+import { DefaultSeo } from 'next-seo';
+
+const favicon = () => {
+  return (
+    <>
+      <link rel="apple-touch-icon" sizes="57x57" href="/favicons/apple-icon-57x57.png" />
+      <link rel="apple-touch-icon" sizes="60x60" href="/favicons/apple-icon-60x60.png" />
+      <link rel="apple-touch-icon" sizes="72x72" href="/favicons/apple-icon-72x72.png" />
+      <link rel="apple-touch-icon" sizes="76x76" href="/favicons/apple-icon-76x76.png" />
+      <link rel="apple-touch-icon" sizes="114x114" href="/favicons/apple-icon-114x114.png" />
+      <link rel="apple-touch-icon" sizes="120x120" href="/favicons/apple-icon-120x120.png" />
+      <link rel="apple-touch-icon" sizes="144x144" href="/favicons/apple-icon-144x144.png" />
+      <link rel="apple-touch-icon" sizes="152x152" href="/favicons/apple-icon-152x152.png" />
+      <link rel="apple-touch-icon" sizes="180x180" href="/favicons/apple-icon-180x180.png" />
+      <link rel="icon" type="image/png" sizes="192x192" href="/favicons/android-icon-192x192.png" />
+      <link rel="icon" type="image/png" sizes="32x32" href="/favicons/favicon-32x32.png" />
+      <link rel="icon" type="image/png" sizes="96x96" href="/favicons/favicon-96x96.png" />
+      <link rel="icon" type="image/png" sizes="16x16" href="/favicons/favicon-16x16.png" />
+      <link rel="manifest" href="/favicons/manifest.json" />
+      <meta name="msapplication-TileColor" content="#ffffff" />
+      <meta name="msapplication-TileImage" content="/favicons/ms-icon-144x144.png" />
+      <meta name="theme-color" content="#ffffff"></meta>
+      <link rel="preconnect" href="https://fonts.googleapis.com" />
+      <link rel="preconnect" href="https://fonts.gstatic.com" crossOrigin="anonymous" />
+      <link
+        href="https://fonts.googleapis.com/css2?family=Fira+Code&family=Nanum+Myeongjo&family=Noto+Sans+KR&display=swap"
+        rel="stylesheet"
+      />
+      <link
+        rel="stylesheet"
+        href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.2.0/styles/a11y-dark.min.css"
+      ></link>
+    </>
+  );
+};
+
+const googleConsole = () => {
+  return (
+    <script
+      async
+      src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=ca-pub-3241811705564924"
+      crossOrigin="anonymous"
+    ></script>
+  );
+};
+
+const defaultSEO = () => {
+  return (
+    <DefaultSeo
+      title={'Kormelon Devlog'}
+      description={`가끔 생각날 적, 정리하고플 적 끄적이는 창고`}
+      canonical="가끔 생각날 때 정리하는 블로그"
+      openGraph={{
+        article: {
+          authors: ['김창회'],
+        },
+        title: 'Kormelon Devlog',
+        description: '가끔 생각날 때 정리하는 블로그',
+        type: 'blog',
+        locale: 'ko_KR',
+        url: 'https://www.kormelon.com/',
+        site_name: 'Kormelon Devlog',
+        images: [
+          {
+            url: `https://images.unsplash.com/photo-1616812757130-aca5451b0243?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=1050&q=80`,
+            alt: `Kormelon Devlog`,
+          },
+        ],
+      }}
+      twitter={{
+        handle: '@handle',
+        site: '@site',
+        cardType: 'summary_large_image',
+      }}
+    />
+  );
+};
+
+export default {
+  favicon,
+  googleConsole,
+  defaultSEO,
+};
